AUTHENTICATION COMPLETED

+++ src/api/routes.py
@@ -17,8 +17,14 @@ jwt = JWTManager()
 # Allow CORS requests to this API
 CORS(api)
 
-# ENDPOINTS DE EVENTOS
-# READ EVENTS CUALQUIERA PUEDE INGRESAR
+#########################################################################################################
+#########################################################################################################
+#########################################################################################################
+# ENDPOINTS DE  LECTURA DE EVENTOS
+# CUALQUIERA PUEDE INGRESAR
+#########################################################################################################
+#########################################################################################################
+#########################################################################################################
 @api.route("/events", methods=["GET"])  
 def handle_events():
     try:
@@ -41,31 +47,42 @@ def read_event(event_id):
         return jsonify({"error": f"{e}"}), 400
 
 
-
-# ENDPOINTS DE USERS CAMBIO 1  ,CAMIBE DE POSICION EL CREATE USER AL PRIMERO EN VEZ DEL 2DO O 3ERO YA QUE ES EL PRIMERO QUE SE VA  A USAR
+##############################################################################################
+##############################################################################################
+##############################################################################################
 # CREATE USER
 @api.route("/users", methods=["POST"])
 def create_user():
     try:
         body = request.get_json()
-        if not body.get("dni") or body.get("name") or body.get("email"):
-            return jsonify({"message": "Error all inputs has to be completed"}),200
+        filter_user = User.query.filter_by(email=body.get("email")).first()
+        
+        if filter_user:
+            return jsonify({"message": "User already exists"}), 400
+        
+        if  not body.get("name") or not body.get("email") or not body.get("password"):
+            return jsonify({"message": "Error, all inputs must be completed"}), 400
+        
         new_user = User(
             dni=body.get("dni"),
             name=body.get("name"),
             last_name=body.get("last_name"),
             email=body.get("email"),
-            password=bcrypt.generate_password_hash(body.get("password")).decode('utf-8'), 
+            password=bcrypt.generate_password_hash(body.get("password")).decode('utf-8'),
             district=body.get("district"),
             phone=body.get("phone"),
             date_of_birth=datetime.fromisoformat(body.get("date_of_birth")),
         )
+        
         db.session.add(new_user)
         db.session.commit()
-        return jsonify({"message": "New User Created", "user": f"{new_user.serialize()}"}),200
+        
+        return jsonify({"message": "New User Created", "user": new_user.serialize()}), 200
+
     except Exception as e:
         db.session.rollback()
         return jsonify({"error": str(e)}), 400
+
 
 #Login de usuario ,paso 2,esto es despues de crearse una cuenta o cuando se vuelva a conectar,probado
 @api.route("/login", methods=["POST"])
@@ -105,7 +122,7 @@ def log_in_user():
     except Exception as e:
         return jsonify({"message": str(e)}), 500
 
-#Este endpoint solo ingresara el user para poder ver sus datos
+#Este endpoint solo ingresara el user para poder ver sus datos estando logeado y nadie más pueda buscar su usuario por ID
 @api.route("/user/", methods=["GET"])
 @jwt_required()
 def read_user_data():
@@ -114,8 +131,11 @@ def read_user_data():
         user = User.query.get(current_user)
         if user:
             return jsonify({"user_data": user.serialize()}), 200
+        else:
+            return jsonify({"error": "User not found"}), 404
     except Exception as e:
         return jsonify({"error": str(e)}), 400
+
 
 #############################################################333
 ###############################################################3
@@ -157,6 +177,8 @@ def handle_users():
         if administrator:
             users = list(map(lambda event: event.serialize(), User.query.all()))
             return jsonify(users), 200
+        else:
+            return jsonify({"error": "Unauthorized"}), 400
     except Exception as e:
         return jsonify({"error": str(e)}), 400
 
@@ -169,16 +191,17 @@ def read_user(user_id):
         user = User.query.get(user_id)
         current_admin = get_jwt_identity()
         administrator=Administrator.query.get(current_admin)
-        if user and administrator:
-            return jsonify(user.serialize()), 200
+        if administrator:
+            if user:
+                return jsonify(user.serialize()), 200
+            else:
+                return jsonify({"error": "User not found"}), 404
         else:
-            return jsonify({"error": "User not found"}), 404
+            return jsonify({"error": "Unauthorized"}), 403   
     except Exception as e:
         return jsonify({"error": str(e)}), 400
 
 ##VALIDADO
-
-
 # UPDATE USER ADMINISTRADOR y USUARIO LOGEADO NADA MÁS PUEDE INGRESAR
 @api.route("/users/<int:user_id>", methods=["PUT"])
 @jwt_required()
@@ -188,17 +211,20 @@ def update_user(user_id):
         user = User.query.get(user_id)
         current_admin = get_jwt_identity()
         administrator=Administrator.query.get(current_admin)
-        if user and administrator:
-            user.dni = body.get("dni", user.dni)
-            user.name = body.get("name", user.name)
-            user.last_name = body.get("last_name", user.last_name)
-            user.email = body.get("email", user.email)
-            user.password = body.get("password", user.password)
-            user.district = body.get("district", user.district)
-            user.phone = body.get("phone", user.phone)
-            user.date_of_birth = datetime.fromisoformat(body.get("date_of_birth"))
-            db.session.commit()
+        if administrator:
+            if user:
+                user.dni = body.get("dni", user.dni)
+                user.name = body.get("name", user.name)
+                user.last_name = body.get("last_name", user.last_name)
+                user.email = body.get("email", user.email)
+                user.password = body.get("password", user.password)
+                user.district = body.get("district", user.district)
+                user.phone = body.get("phone", user.phone)
+                user.date_of_birth = datetime.fromisoformat(body.get("date_of_birth"))
+                db.session.commit()
             return jsonify(user.serialize()), 200
+        else:
+            return jsonify({"error": "Unauthorized"}), 403
     except Exception as e:
         db.session.rollback()
         return jsonify({"error": str(e)}), 400
@@ -209,37 +235,55 @@ def update_user(user_id):
 @jwt_required()
 def delete_user(user_id):
     try:
-        user = User.query.get(user_id)
-        if user:
-            db.session.delete(user)
-            db.session.commit()
-            return jsonify({"message": "User deleted successfully"}), 200
+        current_admin = get_jwt_identity()
+        administrator = Administrator.query.get(current_admin)
+        if administrator:
+            user = User.query.get(user_id)
+            if user:
+                db.session.delete(user)
+                db.session.commit()
+                return jsonify({"message": "User deleted successfully"}), 200
+            else:
+                return jsonify({"error": "User not found"}), 404
         else:
-            return jsonify({"error": "User not found"}), 404
+            return jsonify({"error": "Unauthorized"}), 403
+
     except Exception as e:
         db.session.rollback()
         return jsonify({"error": str(e)}), 400
 
 
+###VALIDADOS"""2
 # CREATE EVENTS ADMINISTRADOR NADA MÁS PUEDE INGRESAR
 @api.route("/events", methods=["POST"])
 @jwt_required()
 def create_event():
     try:
-        body = request.get_json()
-        new_event = Event(
-            name=body.get("name"),
-            date=datetime.fromisoformat(body.get("date")),
-            image_url=body.get("image_url"),
-            place=body.get("place"),
-            description=body.get("description"),
-            category=body.get("category"),
-            stock=body.get("stock"),
-            admin_id=body.get("admin_id"),
-        )
-        db.session.add(new_event)
-        db.session.commit()
-        return jsonify({"message": "New Event Created", "event": f"{new_event.serialize()}"}),200
+        current_admin = get_jwt_identity()
+        administrator = Administrator.query.get(current_admin)
+        if administrator:
+            body = request.get_json()
+            if (body.get("name") and body.get("date") and body.get("image_url") and body.get("place") and body.get("description") and body.get("category") and body.get("stock") and body.get("admin_id")):
+                event_date = datetime.fromisoformat(body.get("date"))
+                new_event = Event(
+                    name=body.get("name"),
+                    date=event_date,
+                    image_url=body.get("image_url"),
+                    place=body.get("place"),
+                    description=body.get("description"),
+                    category=body.get("category"),
+                    stock=body.get("stock"),
+                    admin_id=body.get("admin_id"),
+                )
+                
+                db.session.add(new_event)
+                db.session.commit()
+                
+                return jsonify({"message": "New Event Created", "event": new_event.serialize()}), 200
+            else:
+                return jsonify({"error": "Missing required fields"}), 400
+        else:
+            return jsonify({"error": "Unauthorized"}), 403
     except Exception as e:
         db.session.rollback()
         return jsonify({"error": str(e)}), 400
@@ -250,13 +294,14 @@ def create_event():
 @jwt_required()
 def update_event(event_id):
     try:
-        body = request.get_json()
-        event = Event.query.get(event_id)
-        if event:
+        current_admin = get_jwt_identity()
+        administrator = Administrator.query.get(current_admin)
+        if administrator:
+         body = request.get_json()
+         event = Event.query.get(event_id)
+         if event:
             event.name = body.get("name", event.name)
-            event.date = datetime.fromisoformat(
-                body.get("date", event.date.isoformat())
-            )
+            event.date = datetime.fromisoformat(body.get("date", event.date.isoformat()))
             event.image_url = body.get("image_url", event.image_url)
             event.place = body.get("place", event.place)
             event.description = body.get("description", event.description)
@@ -265,6 +310,9 @@ def update_event(event_id):
             event.admin_id = body.get("admin_id", event.admin_id)
             db.session.commit()
             return jsonify(event.serialize()), 200
+        else:
+            return jsonify({"error": "Unauthorized"}), 403
+
     except Exception as e:
         db.session.rollback()
         return jsonify({"error": str(e)}), 400
@@ -275,11 +323,16 @@ def update_event(event_id):
 @jwt_required()
 def delete_event(event_id):
     try:
-        event = Event.query.get(event_id)
-        if event:
-            db.session.delete(event)
-            db.session.commit()
+        current_admin = get_jwt_identity()
+        administrator = Administrator.query.get(current_admin)
+        if administrator:
+            event = Event.query.get(event_id)
+            if event:
+             db.session.delete(event)
+             db.session.commit()
             return jsonify({"message": "Event deleted successfully"}), 200
+        else:
+            return jsonify({"error": "Unauthorized"}), 403
     except Exception as e:
         db.session.rollback()
         return jsonify({"error": str(e)}), 400
